backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]
passlib

python-multipart
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain-text-splitters
faiss-cpu
youtube-transcript-api
pypdf
beautifulsoup4
requests
sentence-transformers

